+++ src/initialize/index.js
@@ -1,5 +1,5 @@
 import { createDataFile } from './createDataFile.js';
-import { createDotEnv } from './dotEnv.js';
+import { createDotEnv } from './createDotEnv.js';
 import { createGitIgnore } from './createGitIgnore.js';
 import { createPackageJson } from './createPackageJson.js';
 import { createReadme } from './createReadme.js';
